requirements.txt
discord
python-dotenv
fastapi
jinja2
python-multipart
uvicorn[standard]
openai
arrow
rake_nltk